Refactored all GetAimTransform() related functions out into the `AimTransform` class, making `SnapBuilder` less of a `static` god-class.
