+++ SnapBuilder/SnapBuilder.cs
@@ -58,288 +58,6 @@ namespace Straitjacket.Subnautica.Mods.SnapBuilder
             Logger.LogInfo($"Harmony patches applied in {stopwatch.ElapsedMilliseconds}ms.");
         }
 
-        /// <summary>
-        /// Where possible, use the transform of the parent as this should be a better reference point for localisation
-        /// (especially useful inside a base)
-        /// </summary>
-        /// <param name="hit"></param>
-        /// <returns></returns>
-        private static Transform GetAppropriateTransform(RaycastHit hit) => Builder.GetSurfaceType(hit.normal) switch
-        {
-            SurfaceType.Ground => hit.transform.parent ?? hit.transform,
-            _ => hit.transform
-        };
-
-        /// <summary>
-        /// Get a new hit where the point and normal are localised the given transform
-        /// </summary>
-        /// <param name="hit"></param>
-        /// <param name="transform"></param>
-        /// <returns></returns>
-        private static RaycastHit GetLocalisedHit(RaycastHit hit, Transform transform = null)
-        {
-            transform ??= hit.transform;
-            hit.point = transform.InverseTransformPoint(hit.point); // Get the hit point localised relative to the hit transform
-            hit.normal = transform.InverseTransformDirection(hit.normal).normalized; // Get the hit normal localised to the hit transform
-            return hit;
-        }
-
-        /// <summary>
-        /// Gets a new hit in world space
-        /// </summary>
-        /// <param name="hit"></param>
-        /// <param name="transform"></param>
-        /// <returns></returns>
-        private static RaycastHit GetWorldSpaceHit(RaycastHit hit, Transform transform = null)
-        {
-            transform ??= hit.transform;
-            hit.point = transform.TransformPoint(hit.point);
-            hit.normal = transform.TransformDirection(hit.normal).normalized;
-            return hit;
-        }
-
-        /// <summary>
-        /// Gets a new hit where the point is snapped based on the normal and current round factor
-        /// </summary>
-        /// <param name="hitPoint"></param>
-        /// <param name="hitNormal"></param>
-        /// <returns></returns>
-        private static RaycastHit GetSnappedHit(RaycastHit hit)
-        {
-            Vector3 hitPoint = hit.point;
-            Vector3 hitNormal = hit.normal;
-
-            hitNormal.x = Mathf.Abs(hitNormal.x);
-            hitNormal.y = Mathf.Abs(hitNormal.y);
-            hitNormal.z = Mathf.Abs(hitNormal.z);
-            hitNormal = hitNormal.normalized; // For sanity's sake, make sure the normal is normalised
-
-            // Get the rounding factor from user options based on whether the fine snapping key is held or not
-            float roundFactor = Config.FineSnapping.Enabled ? Config.FineSnapRounding / 2f : Config.SnapRounding;
-
-            // Round (snap) the localised hit point coords only on axes where the corresponding normal axis is less than 1
-            if (hitNormal.x < 1)
-            {
-                hitPoint.x = Math.RoundToNearest(hitPoint.x, roundFactor);
-            }
-            if (hitNormal.y < 1)
-            {
-                hitPoint.y = Math.RoundToNearest(hitPoint.y, roundFactor);
-            }
-            if (hitNormal.z < 1)
-            {
-                hitPoint.z = Math.RoundToNearest(hitPoint.z, roundFactor);
-            }
-
-            hit.point = hitPoint;
-            return hit;
-        }
-
-        /// <summary>
-        /// Gets a new hit popped onto the most appropriate surface at the most appropriate point,
-        /// or the original hit if the operation is not possible
-        /// </summary>
-        /// <param name="hit"></param>
-        /// <returns></returns>
-        private static RaycastHit PopHitOntoBestSurface(RaycastHit hit)
-        {
-            if (!Player.main.IsInsideWalkable())
-                return hit;
-
-            switch (Builder.GetSurfaceType(hit.normal))
-            {
-                case SurfaceType.Wall
-                when !Builder.allowedSurfaceTypes.Contains(SurfaceType.Wall)
-                     && Builder.allowedSurfaceTypes.Contains(SurfaceType.Ground):
-
-                    // Get the rotation of the object
-                    Quaternion rotation = Builder.rotationEnabled
-                        ? CalculateRotation(ref Builder.additiveRotation, hit, Builder.forceUpright || Player.main.IsInsideWalkable())
-                        : Quaternion.identity;
-
-                    if (!Builder.bounds.Any())
-                    {
-                        return hit; // if there are no bounds for some reason, just use the original hit
-                    }
-
-                    // Get the corners of the object based on the Builder.bounds, localised to the hit point
-                    IEnumerable<Vector3> corners = Builder.bounds
-                        .Select(bounds => new { Bounds = bounds, Corners = bounds.GetCorners() })
-                        .SelectMany(x => x.Corners.Select(corner => hit.point + rotation * corner));
-
-                    // Get the farthest corner from the player
-                    Vector3 farthestCorner = corners.OrderByDescending(x
-                        => Vector3.Distance(x, AimTransform.Main.OffsetAimTransform.position)).First();
-
-                    // Center the corner to the hit.point on the local X and Y axes
-                    var empty = new GameObject();
-                    var child = new GameObject();
-                    empty.transform.position = hit.point;
-                    empty.transform.forward = hit.normal;
-                    child.transform.SetParent(empty.transform);
-                    child.transform.position = farthestCorner;
-                    child.transform.localPosition = new Vector3(0, 0, child.transform.localPosition.z);
-                    Vector3 farthestCornerCentered = child.transform.position;
-
-                    // Clean up the GameObjects as we don't need them anymore
-                    GameObject.Destroy(child);
-                    GameObject.Destroy(empty);
-
-                    float offset
-#if SUBNAUTICA
-                        = 0.1f; // in subnautica, the collision boundary between objects is much larger than BZ
-#elif BELOWZERO
-                        = 0.01f;
-#endif
-
-                    // Now move the hit.point outward from the wall just enough so that the object can fit
-                    Vector3 poppedPoint = hit.point + hit.normal * Vector3.Distance(farthestCornerCentered, hit.point) + hit.normal * offset;
-
-                    // Try to get a new hit by aiming at the floor from this popped point
-                    if (Raycast(poppedPoint, Vector3.down, out RaycastHit poppedHit))
-                    {
-                        return poppedHit;
-                    }
-
-                    break;
-            }
-
-            return hit;
-        }
-
-        private static int lastCalculationFrame;
-        public static Transform GetAimTransform()
-        {
-            if (!Config.Snapping.Enabled)
-            {
-                return AimTransform.Main.BuilderAimTransform;
-            }
-
-            // Skip recalculating multiple times per frame
-            if (lastCalculationFrame == Time.frameCount)
-            {
-                return AimTransform.Main.transform;
-            }
-            lastCalculationFrame = Time.frameCount;
-
-            // If no hit, exit early
-            if (!Raycast(AimTransform.Main.OffsetAimTransform.position,
-                         AimTransform.Main.BuilderAimTransform.forward,
-                         out RaycastHit hit))
-            {
-                AimTransform.Main.transform.position = AimTransform.Main.OffsetAimTransform.position;
-                AimTransform.Main.transform.forward = AimTransform.Main.BuilderAimTransform.forward;
-                return AimTransform.Main.transform;
-            }
-
-            if (ColliderCache.Main.GetRecord(hit.collider) is ColliderRecord record && record.IsImprovable)
-            {
-                if (Config.DetailedCollider.Enabled && !record.IsImproved)
-                {
-                    ImproveRecordAndUpdateHit(record, ref hit);
-                }
-                else if (!Config.DetailedCollider.Enabled && record.IsImproved)
-                {
-                    RevertRecordAndUpdateHit(record, ref hit);
-                }
-
-                if (hit.collider is null)
-                {
-                    AimTransform.Main.transform.position = AimTransform.Main.OffsetAimTransform.position;
-                    AimTransform.Main.transform.forward = AimTransform.Main.BuilderAimTransform.forward;
-                    return AimTransform.Main.transform;
-                }
-
-                if (Config.RenderImprovableColliders)
-                {
-                    record.Render();
-                }
-            }
-
-            Transform hitTransform = GetAppropriateTransform(hit);
-            RaycastHit localisedHit = GetLocalisedHit(hit, hitTransform);
-            RaycastHit snappedHit = GetSnappedHit(localisedHit);
-            RaycastHit snappedWorldSpaceHit = GetWorldSpaceHit(snappedHit, hitTransform);
-            RaycastHit poppedHit = PopHitOntoBestSurface(snappedWorldSpaceHit);
-
-            AimTransform.Main.transform.position = AimTransform.Main.OffsetAimTransform.position;
-            AimTransform.Main.transform.forward = poppedHit.point - AimTransform.Main.transform.position;
-
-            return AimTransform.Main.transform;
-        }
-
-        private static void ImproveRecordAndUpdateHit(ColliderRecord record, ref RaycastHit hit)
-        {
-            record.Improve();
-            if (record.IsImproved)
-            {
-                Raycast(AimTransform.Main.OffsetAimTransform.position,
-                          AimTransform.Main.BuilderAimTransform.forward,
-                          out hit);
-            }
-        }
-
-        private static void RevertRecordAndUpdateHit(ColliderRecord record, ref RaycastHit hit)
-        {
-            record.Revert();
-            if (!record.IsImproved)
-            {
-                Raycast(AimTransform.Main.OffsetAimTransform.position,
-                          AimTransform.Main.BuilderAimTransform.forward,
-                          out hit);
-            }
-        }
-
-        private static bool Raycast(Vector3 from, Vector3 direction, out RaycastHit hit) =>
-            Physics.Raycast(from, direction, out hit, Builder.placeMaxDistance,
-                Builder.placeLayerMask, QueryTriggerInteraction.Ignore);
-
-        private static void ImproveHitNormal(ref RaycastHit hit)
-        {
-            // If the hit.collider is a MeshCollider and has a sharedMesh, it is a surface like the ground or the roof of a multipurpose room,
-            // in which case we want a more accurate normal where possible
-            if (hit.collider is MeshCollider meshCollider && meshCollider.sharedMesh is Mesh)
-            {
-                // Set up the offsets for raycasts around the point
-                Vector3[] offsets = new Vector3[]
-                {
-                    Vector3.forward,
-                    Vector3.back,
-                    Vector3.left,
-                    Vector3.right,
-                    (Vector3.forward + Vector3.right) * .707f,
-                    (Vector3.forward + Vector3.left) * .707f,
-                    (Vector3.back + Vector3.right) * .707f,
-                    (Vector3.back + Vector3.left) * .707f
-                };
-
-                List<Vector3> normals = new List<Vector3>();
-                // Perform a raycast from each offset, pointing down toward the surface
-                foreach (Vector3 offset in offsets)
-                {
-                    Raycast(hit.point + Vector3.up * .1f + offset * .2f,
-                            Vector3.down,
-                            out RaycastHit offsetHit);
-                    if (offsetHit.transform == hit.transform) // If we hit the same object, add the hit normal
-                    {
-                        normals.Add(offsetHit.normal);
-                    }
-                }
-
-                foreach (Vector3 normal in normals)
-                {
-                    if (!normal.Equals(hit.normal))
-                    {   // If the normal is not the same as the original, add it to the normal and average
-                        hit.normal += normal;
-                        hit.normal /= 2;
-                    }
-                }
-
-                // For sanity's sake, make sure the normal is normalised after summing & averaging
-                hit.normal = hit.normal.normalized;
-            }
-        }
-
         private static void ApplyAdditiveRotation(ref float additiveRotation)
         {
             // If the user is rotating, apply the additive rotation
@@ -385,9 +103,9 @@ namespace Straitjacket.Subnautica.Mods.SnapBuilder
         public static Quaternion CalculateRotation(ref float additiveRotation, RaycastHit hit, bool forceUpright)
         {
             ApplyAdditiveRotation(ref additiveRotation);
-            ImproveHitNormal(ref hit);
+            hit.normal = hit.AverageNormal(.2f);
 
-            Transform hitTransform = GetAppropriateTransform(hit);
+            Transform hitTransform = hit.GetOptimalTransform();
 
             // Instantiate empty game objects for applying rotations
             GameObject empty = new GameObject();
@@ -423,8 +141,8 @@ namespace Straitjacket.Subnautica.Mods.SnapBuilder
             Quaternion rotation = child.transform.rotation; // Our final rotation
 
             // Clean up after ourselves
-            GameObject.DestroyImmediate(child);
-            GameObject.DestroyImmediate(empty);
+            Object.DestroyImmediate(child);
+            Object.DestroyImmediate(empty);
 
             return rotation;
         }
